Backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
pandas==3.0.2
scikit-learn==1.7.2
joblib==1.4.2
python-dotenv==1.0.1
